# apps/mrerp/backend/openapi.yaml
openapi: 3.0.3
info:
  title: MRERP API
  version: 1.0.0
  description: MRERP modular-monolith API contract
paths:
...
  /api/v1/rewards/allowances/:
    get:
      operationId: api_v1_rewards_allowances_retrieve
      tags:
      - api
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
    post:
      operationId: api_v1_rewards_allowances_create
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AllowanceInput'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AllowanceInput'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AllowanceInput'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
  /api/v1/rewards/gifts/:
    get:
      operationId: api_v1_rewards_gifts_list
      tags:
      - api
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Gift'
          description: ''
    post:
      operationId: api_v1_rewards_gifts_create
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Gift'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Gift'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Gift'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gift'
          description: ''
  /api/v1/rewards/gifts/{uuid}/:
    patch:
      operationId: api_v1_rewards_gifts_partial_update
      parameters:
      - in: path
        name: uuid
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedGiftPatch'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedGiftPatch'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedGiftPatch'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gift'
          description: ''
  /api/v1/rewards/leaderboard/:
    get:
      operationId: api_v1_rewards_leaderboard_list
      parameters:
      - in: query
        name: period
        schema:
          type: string
          enum:
          - month
          - quarter
          - year
          default: month
      tags:
      - api
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaderboardRow'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: ''
...
  /api/v1/rewards/redemptions/:
    get:
      operationId: api_v1_rewards_redemptions_retrieve
      tags:
      - api
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedemptionPage'
          description: ''
    post:
      operationId: api_v1_rewards_redemptions_create
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RedeemInput'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RedeemInput'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RedeemInput'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Redemption'
          description: ''
  /api/v1/rewards/redemptions/{uuid}/decision/:
    post:
      operationId: api_v1_rewards_redemptions_decision_create
      parameters:
      - in: path
        name: uuid
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionInput'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/DecisionInput'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DecisionInput'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Redemption'
          description: ''
  /api/v1/rewards/stars/grant/:
    post:
      operationId: api_v1_rewards_stars_grant_create
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StarEntryCreate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/StarEntryCreate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/StarEntryCreate'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StarEntry'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: ''
  /api/v1/rewards/stars/me/:
    get:
      operationId: api_v1_rewards_stars_me_retrieve
      tags:
      - api
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Balance'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: ''
...
components:
  schemas:
...
    AllowanceInput:
      type: object
      properties:
        team_uuid:
          type: string
          format: uuid
        month:
          type: string
          format: date
        limit:
          type: integer
          maximum: 1000000
          minimum: 0
      required:
      - limit
      - month
      - team_uuid
    ApiError:
      type: object
      properties:
        code:
          type: string
        detail:
          type: string
        correlation_id:
          type: string
          format: uuid
        errors:
          type: object
          additionalProperties: {}
      required:
      - code
      - correlation_id
      - detail
...
    Balance:
      type: object
      properties:
        held:
          type: integer
        balance:
          type: integer
        ledger:
          type: array
          items:
            $ref: '#/components/schemas/StarEntry'
      required:
      - balance
      - held
      - ledger
...
    DecisionInput:
      type: object
      properties:
        action:
          $ref: '#/components/schemas/DecisionInputActionEnum'
        note:
          type: string
          maxLength: 500
      required:
      - action
    DecisionInputActionEnum:
      enum:
      - approve
      - reject
      - cancel
      - fulfill
      type: string
      description: |-
        * `approve` - approve
        * `reject` - reject
        * `cancel` - cancel
        * `fulfill` - fulfill
...
    EntryTypeEnum:
      enum:
      - grant
      - adjustment
      - hold
      - refund
      type: string
      description: |-
        * `grant` - Cấp sao
        * `adjustment` - Điều chỉnh
        * `hold` - Giữ sao đổi quà
        * `refund` - Hoàn sao đổi quà
...
    Gift:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
          readOnly: true
        title:
          type: string
          maxLength: 180
        description:
          type: string
          maxLength: 3000
        category:
          type: string
          maxLength: 80
        cost:
          type: integer
          maximum: 1000000
          minimum: 1
        stock:
          type: integer
          maximum: 1000000
          minimum: 0
        active:
          type: boolean
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - category
      - cost
      - stock
      - title
      - updated_at
      - uuid
...
    LeaderboardRow:
      type: object
      properties:
        rank:
          type: integer
        employee_uuid:
          type: string
          format: uuid
        display_name:
          type: string
        employee_code:
          type: string
        team_name:
          type: string
          nullable: true
        stars:
          type: integer
      required:
      - display_name
      - employee_code
      - employee_uuid
      - rank
      - stars
      - team_name
...
    PatchedGiftPatch:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
          readOnly: true
        title:
          type: string
          maxLength: 180
        description:
          type: string
          maxLength: 3000
        category:
          type: string
          maxLength: 80
        cost:
          type: integer
          maximum: 1000000
          minimum: 1
        stock:
          type: integer
          maximum: 1000000
          minimum: 0
        active:
          type: boolean
        updated_at:
          type: string
...
    RedeemInput:
      type: object
      properties:
        gift_uuid:
          type: string
          format: uuid
        request_key:
          type: string
          format: uuid
        expected_cost:
          type: integer
          minimum: 1
      required:
      - expected_cost
      - gift_uuid
      - request_key
    Redemption:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
          readOnly: true
        employee:
          type: string
          format: uuid
          readOnly: true
        employee_name:
          type: string
          readOnly: true
        gift:
          type: string
          format: uuid
          readOnly: true
        gift_title:
          type: string
          readOnly: true
        cost:
          type: integer
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/RedemptionStatusEnum'
          readOnly: true
        status_label:
          type: string
          readOnly: true
        note:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - cost
      - created_at
      - employee
      - employee_name
      - gift
      - gift_title
      - note
      - status
      - status_label
      - updated_at
      - uuid
    RedemptionPage:
      type: object
      properties:
        count:
          type: integer
        next:
          type: string
          nullable: true
        previous:
          type: string
          nullable: true
        results:
          type: array
          items:
            $ref: '#/components/schemas/Redemption'
      required:
      - count
      - next
      - previous
      - results
    RedemptionStatusEnum:
      enum:
      - pending
      - approved
      - fulfilled
      - rejected
      - cancelled
      type: string
      description: |-
        * `pending` - Chờ duyệt
        * `approved` - Đã duyệt
        * `fulfilled` - Đã trao quà
        * `rejected` - Từ chối
        * `cancelled` - Đã hủy
...
    StarEntry:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
          readOnly: true
        employee:
          type: string
          format: uuid
          readOnly: true
        employee_name:
          type: string
          readOnly: true
        amount:
          type: integer
          readOnly: true
        entry_type:
          allOf:
          - $ref: '#/components/schemas/EntryTypeEnum'
          readOnly: true
        entry_type_label:
          type: string
          readOnly: true
        reason:
          type: string
          readOnly: true
        actor_name:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - actor_name
      - amount
      - created_at
      - employee
      - employee_name
      - entry_type
      - entry_type_label
      - reason
      - uuid
    StarEntryCreate:
      type: object
      properties:
        employee_uuid:
          type: string
          format: uuid
        amount:
          type: integer
          maximum: 1000000
          minimum: -1000000
        reason:
          type: string
          maxLength: 500
          minLength: 1
        idempotency_key:
          type: string
          maxLength: 120
      required:
      - amount
      - employee_uuid
      - reason
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid